Update Promocion.java
cambie otro int por double, voy despacio porque no quiero hacer macana

+++ src/tierra_Media/Promocion.java
@@ -2,7 +2,7 @@ package tierra_Media;
 
 public abstract class Promocion {
 	Atraccion[] atracciones;
-	int costoDeVisita;
+	double costoDeVisita;
 	public Promocion() {
 		
 	}
